requirements.txt
requests==2.27.1
discord.py==2.3.2